update user profile page; add image to show

+++ Tryit/Tryit/UI/ProfileViewController.m
@@ -69,7 +69,7 @@ NSString *const UserRestCellIdentifier = @"UserRestCellIdentifier";
     [UIFunction showWaitingAlertWithString:NSLocalizedString(@"PROMPT_LODING", nil)];
     [WebAPI getRestaurantsSuccess:^(NSMutableArray *array) {
         weakSelf_SC.restArray = [NSMutableArray arrayWithArray:array];
-        self.userProfile.vipRest = [NSString stringWithFormat:@"%d", array.count];
+        self.userProfile.vipRest = [NSString stringWithFormat:@"%lu", (unsigned long)array.count];
         ProfileInfoView *profileView = (ProfileInfoView*)self.tableView.tableHeaderView;
         profileView.userProfile = self.userProfile;
         [weakSelf_SC.tableView reloadData];
